fastapi/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
passlib[bcrypt]
pyjwt
python-jose[cryptography]